cmd: test: Add bitwise AND, document the feature
Currently, there is not a good way to control script flow based on
bitwise values. This initially led to NI adding the ability to read
a specific bit from a port-mapped register to control bootflow.

Here, we add a more generic ability to test the bitwise AND of a
value available to the grub scripting environment. This obviates the
need for the inbit command.

Signed-off-by: Brad Mouring <brad.mouring@ni.com>